Add report generation functionality
This commit adds the functionality to generate reports in the library management system. The reports can be generated for users, books, and transactions. The reports are generated in the form of CSV files and stored in the reports folder. The reports are generated using the data stored in the database using the Slick library. The commit also includes the necessary changes to the Main.scala file to handle the report generation menu options.